layers/common/python/requirements.txt
PyJWT>=2.8.0
bcrypt>=4.1.0
